Incorporate PR feedback. Refactor LLVMInlineAssembly test suite
