+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly001.c
