+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly002.c
