+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly003.c
