+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly004.c
