+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly005.c
