+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly006.c
