+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly007.c
