+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly008.c
