+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly009.c
