+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly010.c
