+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly011.c
