+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly012.c
