+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly014.c
