+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly015.c
