+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly016.c
