+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly017.c
