+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly018.c
