+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly019.c
