+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly020.c
