+++ projects/com.oracle.truffle.llvm.test/inlineassemblytests/inlineassembly021.c
