--- projects/com.oracle.truffle.llvm.test/tests/c/inlineassembly/inlineassembly018.reference_output
@@ -1,1 +1,0 @@
-exit 12
