--- projects/com.oracle.truffle.llvm.test/tests/c/inlineassembly/inlineassembly020.reference_output
@@ -1,1 +1,0 @@
-exit 80
